+++ WireFoundation/Tests/WireUtilitiesPackageTests/FileArchiving/ZIPFoundationFileArchiverTests.swift
@@ -16,19 +16,35 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
+import Foundation
 import Testing
 
-@testable import WireFoundation
+@testable import WireUtilitiesPackage
 
 struct ZIPFoundationFileArchiverTests {
 
-    @Test func todo() async throws {
-        // TODO: finish
-        fatalError()
+    @Test func `test creating an archive from a single file`() async throws {
+        // Given
+        let fileManager = FileManager.default
+        let temporaryDirectory = try fileManager.temporaryDirectory(create: true)
+        defer { try? fileManager.removeItem(at: temporaryDirectory) }
+        let fileURL = temporaryDirectory.appending(path: "C.txt", directoryHint: .notDirectory)
+        let archiveURL = temporaryDirectory.appending(path: "C.zip", directoryHint: .notDirectory)
+        try "-C-".write(to: fileURL, atomically: true, encoding: .utf8)
+        let sut = ZIPFoundationFileArchiver()
+
+        // When
+        try sut.zipResources(at: [fileURL], into: archiveURL)
+
+        // Then
+        let unzippedDirectory = temporaryDirectory.appending(path: "C", directoryHint: .isDirectory)
+        try ZIPFoundationFileUnarchiver().unzipFile(at: archiveURL, to: unzippedDirectory)
+        let content = try String(contentsOf: unzippedDirectory.appending(path: "C.txt", directoryHint: .notDirectory))
+        #expect(content == "-C-")
     }
 
     // TODO: test cleanup, no permission, invalid url, unzipping, non-existent source files
 
 }
 
-// TODO: test wire backup manually
+// TODO: test WireBackup manually
